Feat implement Name value object
